graph - down to 3 tests
