--- tests/graph/unit/test_graph.py
@@ -1,5 +1,0 @@
-from tests.utils import base
-
-
-class TestGraph(base.TestCase):
-    pass
