--- tests/graph/unit/test_query_result.py
@@ -1,5 +1,0 @@
-from tests.utils import base
-
-
-class TestQueryResult(base.TestCase):
-    pass
