+++ tests/test_bloom.py
@@ -8,7 +8,7 @@ i = lambda l: [int(v) for v in l]
 
 @pytest.fixture
 def client():
-    rc = RedisPlus(Redis())  # modules={'bf': {"client": Redis()}})
+    rc = RedisPlus(Redis())
     assert isinstance(rc.bloom, redisplus.bf.Bloom)
     rc.bloom.flushdb()
